simplified the code

+++ src/main.rs
@@ -1,22 +1,23 @@
 use std::fs::File;
 
-use inquire::{Editor, Select, Text};
-use serde::{Deserialize, Serialize};
-use serde_repr::{Deserialize_repr, Serialize_repr};
+use inquire::{Confirm, Editor, Select, Text};
 
-#[derive(Debug, Clone, Default, Deserialize, Serialize)]
-struct Details {
-    pub title: String,
-    pub author: String,
-    pub artist: String,
-    pub description: String,
-    pub genre: Vec<String>,
+use serde::Serialize;
+use serde_repr::Serialize_repr;
+
+#[derive(Debug, Clone, Default, Serialize)]
+pub struct Details<'a> {
+    pub title: &'a str,
+    pub author: &'a str,
+    pub artist: &'a str,
+    pub description: &'a str,
+    pub genre: Vec<&'a str>,
     pub status: Status,
 }
 
-#[derive(Debug, Default, Clone, Copy, Serialize_repr, Deserialize_repr)]
-#[repr(i8)]
-enum Status {
+#[derive(Debug, Default, Clone, Copy, Serialize_repr)]
+#[repr(u8)]
+pub enum Status {
     #[default]
     Unknown = 0,
     Ongoing = 1,
@@ -45,36 +46,37 @@ impl std::fmt::Display for Status {
 }
 
 fn main() -> std::io::Result<()> {
-    let detail = prompt_details();
-    create_file(&detail)
-}
-
-fn prompt_details() -> Details {
     let title = Text::new("Title").prompt().unwrap_or_default();
     let author = Text::new("Author(s)").prompt().unwrap_or_default();
     let artist = Text::new("Artist(s)").prompt().unwrap_or_default();
     let description = Editor::new("Description").prompt().unwrap_or_default();
-    let genre = Text::new("Genre(s)").prompt().unwrap_or_default();
+    let genre = Text::new("Genre(s)")
+        .with_help_message("Genres must be separated with a comma (genre1, genre2, etc)")
+        .prompt()
+        .unwrap_or_default();
     let status = Select::new("Status", Status::VARIANTS.to_vec())
         .prompt()
         .unwrap_or_default();
 
-    Details {
-        title,
-        author,
-        artist,
-        description,
-        genre: genre
-            .split(',')
-            .map(|s| s.trim())
-            .map(|s| s.to_string())
-            .collect(),
+    let detail = Details {
+        title: &title,
+        author: &author,
+        artist: &artist,
+        description: &description,
+        genre: genre.split(',').map(|s| s.trim()).collect(),
         status,
-    }
-}
+    };
 
-fn create_file(details: &Details) -> std::io::Result<()> {
-    let file = File::create("details.json")?;
-    serde_json::to_writer_pretty(file, details)?;
+    if let Ok(true) = Confirm::new("Write to file? [./details.json]")
+        .with_default(false)
+        .prompt()
+    {
+        let file = File::create("details.json")?;
+        serde_json::to_writer_pretty(file, &detail)?;
+        println!("Written to ./details.json")
+    }
+
+    println!("{}", serde_json::to_string_pretty(&detail)?);
+
     Ok(())
 }
